[tdd][springboot] 패키지 정리
