--- tdd/spring-boot/src/main/java/com/example/orderservice/order/OrderPort.java
@@ -1,9 +1,0 @@
-package com.example.orderservice.order;
-
-import com.example.orderservice.product.Product;
-
-public interface OrderPort {
-    Product getProductById(Long productId);
-
-    void save(Order order);
-}
