+++ tdd/spring-boot/src/main/java/com/example/orderservice/order/adapter/OrderAdapter.java
@@ -1,7 +1,10 @@
-package com.example.orderservice.order;
+package com.example.orderservice.order.adapter;
 
-import com.example.orderservice.product.Product;
-import com.example.orderservice.product.ProductRepository;
+import com.example.orderservice.order.application.port.OrderPort;
+import com.example.orderservice.order.domain.Order;
+import com.example.orderservice.order.domain.OrderRepository;
+import com.example.orderservice.product.domain.Product;
+import com.example.orderservice.product.domain.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
