+++ tdd/spring-boot/src/main/java/com/example/orderservice/order/application/service/OrderService.java
@@ -1,6 +1,9 @@
-package com.example.orderservice.order;
+package com.example.orderservice.order.application.service;
 
-import com.example.orderservice.product.Product;
+import com.example.orderservice.order.application.port.OrderPort;
+import com.example.orderservice.order.domain.Order;
+import com.example.orderservice.order.interfaces.dto.CreateOrderRequest;
+import com.example.orderservice.product.domain.Product;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
