+++ tdd/spring-boot/src/main/java/com/example/orderservice/order/domain/OrderRepository.java
@@ -1,5 +1,6 @@
-package com.example.orderservice.order;
+package com.example.orderservice.order.domain;
 
+import com.example.orderservice.order.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
