+++ tdd/spring-boot/src/main/java/com/example/orderservice/order/interfaces/dto/CreateOrderRequest.java
@@ -1,4 +1,4 @@
-package com.example.orderservice.order;
+package com.example.orderservice.order.interfaces.dto;
 
 import lombok.Getter;
