--- tdd/spring-boot/src/main/java/com/example/orderservice/product/DiscountPolicy.java
@@ -1,5 +1,0 @@
-package com.example.orderservice.product;
-
-enum DiscountPolicy {
-    NONE
-}
