--- tdd/spring-boot/src/main/java/com/example/orderservice/product/ProductPort.java
@@ -1,7 +1,0 @@
-package com.example.orderservice.product;
-
-interface ProductPort {
-    void save(Product product);
-
-    Product getProduct(Long productId);
-}
