+++ tdd/spring-boot/src/main/java/com/example/orderservice/product/application/port/ProductPort.java
@@ -1,0 +1,9 @@
+package com.example.orderservice.product.application.port;
+
+import com.example.orderservice.product.domain.Product;
+
+interface ProductPort {
+    void save(Product product);
+
+    Product getProduct(Long productId);
+}
