+++ tdd/spring-boot/src/main/java/com/example/orderservice/product/domain/DiscountPolicy.java
@@ -1,0 +1,5 @@
+package com.example.orderservice.product.domain;
+
+enum DiscountPolicy {
+    NONE
+}
