+++ tdd/spring-boot/src/main/java/com/example/orderservice/product/domain/ProductRepository.java
@@ -1,10 +1,7 @@
-package com.example.orderservice.product;
+package com.example.orderservice.product.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.HashMap;
-import java.util.Map;
 
 @Repository
 public
