+++ tdd/spring-boot/src/test/java/com/example/orderservice/product/ProductTest.java
@@ -1,5 +1,7 @@
 package com.example.orderservice.product;
 
+import com.example.orderservice.product.domain.DiscountPolicy;
+import com.example.orderservice.product.domain.Product;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
